requirements.txt
rpaframework==22.5.3
pydantic==1.10.7
